% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_J}
\alias{build_J}
\title{Build directed indicator matrix}
\usage{
build_J(adj_mat)
}
\arguments{
\item{adj_mat}{The original adjacency matrix.}
}
\value{
A directed indicator matrix \code{J} in sparse form.
}
\description{
Build the sparse directed indicator matrix \code{J}
from a graph adjacency matrix.
}
\keyword{internal}
