% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_J0}
\alias{build_J0}
\title{Build missing-edge indicator matrix}
\usage{
build_J0(adj_mat)
}
\arguments{
\item{adj_mat}{The original adjacency matrix.}
}
\value{
A missing-edge indicator matrix \code{J0}.
}
\description{
Build the missing-edge indicator matrix \code{J0} from a
graph adjacency matrix.
}
\keyword{internal}
