% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_Jd}
\alias{build_Jd}
\title{Build double-edge indicator matrix}
\usage{
build_Jd(adj_mat)
}
\arguments{
\item{adj_mat}{The original adjacency matrix.}
}
\value{
A double-edge indicator matrix \code{Jd} in sparse form.
}
\description{
Build the sparse double-edge indicator matrix \code{Jd} from a
graph adjacency matrix.
}
\keyword{internal}
