% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_Je}
\alias{build_Je}
\title{Build edge-and-diagonal indicator matrix}
\usage{
build_Je(adj_mat)
}
\arguments{
\item{adj_mat}{The original adjacency matrix.}
}
\value{
An edge-and-diagonal indicator matrix \code{Je} in sparse form.
}
\description{
Build the sparse edge-and-diagonal indicator matrix \code{Je} from a
graph adjacency matrix.
}
\keyword{internal}
