% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_Jn}
\alias{build_Jn}
\title{Build vertex-distinct indicator matrix}
\usage{
build_Jn(adj_mat)
}
\arguments{
\item{adj_mat}{The original adjacency matrix.}
}
\value{
A vertex-distinct indicator matrix \code{Jn}.
}
\description{
Build the vertex-distinct indicator matrix \code{Jn} from a
graph adjacency matrix.
}
\keyword{internal}
