% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_Js}
\alias{build_Js}
\title{Build single-edge indicator matrix}
\usage{
build_Js(adj_mat)
}
\arguments{
\item{adj_mat}{The original adjacency matrix.}
}
\value{
A single-edge indicator matrix \code{Js} in sparse form.
}
\description{
Build the sparse single-edge indicator matrix \code{Js} from a
graph adjacency matrix.
}
\keyword{internal}
