% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop0_killdiag}
\alias{drop0_killdiag}
\title{Set diagonal entries to zero and sparsify}
\usage{
drop0_killdiag(some_mat)
}
\arguments{
\item{some_mat}{A square matrix.}
}
\value{
A sparse-form copy of \code{some_mat} with its
diagonal entries set to zero.
}
\description{
Set the diagonal entries of a matrix to zero
and convert it to sparse form.
}
\keyword{internal}
