% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{get_first_eigs}
\alias{get_first_eigs}
\title{Compute first few eigenvalues and eigenvectors of a matrix}
\usage{
get_first_eigs(some_mat, num_eigs)
}
\arguments{
\item{some_mat}{Matrix for which eigenvalues and eigenvectors
are to be calculated.}

\item{num_eigs}{Number of eigenvalues and eigenvectors to calculate.}
}
\value{
A list with two entries:
\code{vals} contains a length-\code{num_eigs} vector of the first few
eigenvalues,
and vects contains an \code{nrow(some_mat)} by \code{num_eigs} matrix
of the associated eigenvectors.
}
\description{
Compute the first few eigenvalues (by magnitude) and
associated eigenvectors of a matrix.
}
\keyword{internal}
