% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{mam_M10}
\alias{mam_M10}
\title{Perform the motif adjacency matrix calculations for motif M10}
\usage{
mam_M10(adj_mat, motif_type, mam_weight_type, mam_method)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the motif
adjacency matrix.}

\item{motif_type}{Type of motif adjacency matrix to build.}

\item{mam_weight_type}{The weighting scheme to use.
One of \code{"unweighted"}, \code{"mean"} or \code{"product"}.}

\item{mam_method}{Which formulation to use.
One of \code{"dense"} or \code{"sparse"}.}
}
\value{
A motif adjacency matrix.
}
\description{
Perform the motif adjacency matrix calculations for motif M10
}
\keyword{internal}
