% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_sparse_matrix}
\alias{random_sparse_matrix}
\title{Build a random sparse matrix}
\usage{
random_sparse_matrix(m, n, p, sample_weight_type = "constant", w = 1)
}
\arguments{
\item{m, n}{Dimension of matrix to build is \code{(m, n)}.}

\item{p}{Probability that each entry is non-zero (before weighting).}

\item{sample_weight_type}{Type of weighting scheme.}

\item{w}{Weight parameter.}
}
\value{
A random sparse matrix.
}
\description{
Build a sparse matrix of size \code{m * n} with
non-zero probability \code{p}.
Edge weights can be unweighted, constant-weighted or
Poisson-weighted.
}
